requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
tqdm>=4.66
requests>=2.31
beautifulsoup4>=4.12
feedparser>=6.0
trafilatura>=1.6          # robust article extraction
newspaper3k>=0.2.8        # optional fallback extractor
spacy>=3.7
textblob>=0.17
scikit-learn>=1.3
umap-learn>=0.5.5
hdbscan>=0.8.36
sentence-transformers>=2.5
transformers>=4.40
torch>=2.1                # CPU is fine; install CUDA build if you have a GPU
python-dotenv>=1.0
yfinance>=0.2.40
openai>=1.30.0
seaborn>=0.13.2
